After much struggling, I figured out how to make the temperatures match up. Here's how I figured it out:
The first place I looked was in the temperature calculation because if something was wrong there it could be a quick fix. I didn't find anything there, so I looked to what the temperature was based off of, and that was the energy. I reasoned that it had to be one of these four things:
    1. The earth gains energy too quickly.
    2. The earth loses energy too slowly.
    3. Temperature is calculated incorrectly.
    4. There is an inaccuracy in keeping track of the earth's energy (precision and rounding errors).
I verified that the rate of energy gain was consistent between the two simulations (mine and PhET's), so it wasn't #1. I saw no way it could be a precision error because the original sim uses Doubles, and floating-point numbers in Javascript are the same size as Java Doubles (64 bits), and there was no explicit rounding anywhere in the code (and none that could be seen when printing out values), so it couldn't be #4.  I had gone over the code a bunch of times to pretty much rule out #3.  So I looked into the possibility that it might be retaining too much energy.  When printing out the wavelength of photons being absorbed by the earth, I noticed that for every sunlight-wavelength photon being absorbed, there was a flag-raising number of infrared photons being absorbed. They were being instantly reabsorbed. After doing some tests to change the requirements for absorption so the reabsorption was reduced to almost nothing (and nothing in the case of the glass-layers tab), I found that the temperatures then matched the original. So here in this fix I've changed where the newly emitted infrared photons are spawned to reduce that reabsorption rate. I don't know if this is the best fix for reabsorption, but I do know definitively now that photon reabsorption is the root of the problem.