Switching workstations. What I have observed is this:
If the triangle's X >= 0, it intersects the right side first;
else, it intersects something that is not quite the left and goes crazy.

I'm not sure what the problem is yet, but I've thoroughly checked for places where the initial incidentRay's directionUnitVector is being changed when it should stay the same and wasn't able to find any instances. (This commit actually makes sure that's true.)