Revert "Trying to better control the order in which recalculations occur after junctions are changed to reduce the errors."
This reverts commit 6f2126b012a6bdf54a116fb13733a593dbfdd348. (Also, I'm choosing not to commit a set of similar-minded changes.)

I think the root of my issue is that the operations I perform on the circuit from the UI are not thread-safe.  I don’t think a good solution would be to make everything happen in the loop, either, because I’d have to make a whole action-queueing system.  I’ve already started to pull some stuff out into the loop like the removal of branches, but it’s only going to get harder from here.  No, I think I need to make sure the events and triggers run as efficiently as possible and just rely on those to get the job done.  What is still an issue, however, is that the particleSet update still happens in the simulation’s update loop, when in fact the objects that it relies upon could be changing concurrently from actions initiated in the UI.